nl2sql/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.11
python-dotenv==1.0.1
google-generativeai==0.8.3
groq
pydantic==2.9.2
jinja2==3.1.4
python-multipart
